+++ inst/examples/readRappOptionFile.r
@@ -1,0 +1,10 @@
+\dontrun{
+
+if (hasRuntimeOptionFile()) {
+  readRappOptionFile()
+}
+if (hasOptionFile()) {
+  readRappOptionFile(path = "rapp/options/options.r")
+}
+
+}