+++ tests/testthat/test-mergeNamespaceRappOptions-1.r
@@ -1,11 +1,11 @@
-context("ensureNamespaceRappOptions-1")
-test_that("ensureNamespaceRappOptions", {
+context("mergeNamespaceRappOptions-1")
+test_that("mergeNamespaceRappOptions", {
 
   opts_old <- getOption(".rapp")
   
   initializeRappOptions()
   setRappGlobal(tempdir())
-  expect_is(res <- ensureNamespaceRappOptions(), "logical")
+  expect_is(res <- mergeNamespaceRappOptions(), "logical")
   expect_equal(ls(getOption(".rapp"), all.names=TRUE), c(".rte", "rapp.core.rte"))
   expect_true(all(c("ns", "ns_global") %in% 
         ls(getOption(".rapp")$rapp.core.rte, all.names=TRUE)))